backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.39",
    "stream-chat>=4.23.0",
]
